refactor: 重構命名方式、新增判斷 gtm click classes
- 命名規則不統一
- config 中 favicon 命名
- contentful CTA model 中新增自訂義 class 的欄位 (for gtm click classes)，程式需判斷有無填寫此欄位
