+++ src/components/Card/BasicCard.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
-import Link from "../link"
+import Link from "../Link"
 import ContinueReading from "./ContinueReading"
 import CategoryAndUpdateTime from "./CategoryAndUpdateTime"
