+++ src/components/Card/CategoryAndUpdateTime.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
-import Link from "../link"
+import Link from "../Link"
 
 const CategoryAndUpdateTime = ({ topicSlug, topicName, updatedAt }) => {
   return (
