+++ src/components/Card/ContinueReading.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
-import Link from "../link"
-import { Button } from "../button"
+import Link from "../Link"
+import { Button } from "../Button"
 import { ArrowRightIcon } from "@heroicons/react/outline"
 
 const ContinueReading = ({ topicSlug, slug }) => {
