+++ src/components/Cta.jsx
@@ -1,17 +1,13 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { Button } from "./button"
-import Link from "./link"
+import { Button } from "./Button"
+import Link from "./Link"
 
 //用於文章頁下方的CTA
-export function CTA({
-  className,
-  destination = "",
-  excerpt = "",
-  title = "立即諮詢",
-}) {
+export function Cta({ data }) {
+  const { title, link, excerpt } = data
   return (
-    <div className={`mx-auto w-full py-16 ${className}`}>
+    <div className="mx-auto w-full py-16">
       <div className="space-y-8 border-y-2 border-primary-500 py-8 sm:py-12">
         {excerpt && (
           <p className="heading text-center text-2xl sm:text-3xl !leading-loose tracking-wide font-bold">
@@ -19,8 +15,10 @@ export function CTA({
           </p>
         )}
         <div className="flex justify-center">
-          <Link as="a" to={destination}>
-            <Button className="contact">{title}</Button>
+          <Link as="a" to={link}>
+            <Button className={`convert ${data?.gtmClickClasses || ""}`}>
+              {title}
+            </Button>
           </Link>
         </div>
       </div>
@@ -28,9 +26,6 @@ export function CTA({
   )
 }
 
-CTA.propTypes = {
-  className: PropTypes.string,
-  destination: PropTypes.string,
-  excerpt: PropTypes.string,
-  title: PropTypes.string,
+Cta.propTypes = {
+  data: PropTypes.string,
 }
