+++ src/components/Index/LatestPost.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
-import Card from "../card/Card"
+import Card from "../Card/Card"
 
 const LatestPost = ({ posts }) => {
   return posts.map(post => {
