+++ src/components/Index/PostCategory.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
-import Link from "../link"
+import Link from "../Link"
 import { GatsbyImage } from "gatsby-plugin-image"
 
 const PostCategory = ({ topics }) => {
