+++ src/components/Post/PopularPost.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
 import { Text } from "../../ui"
-import Card from "../card/Card"
+import Card from "../Card/Card"
 
 const PopularPost = ({ popularPosts }) => {
   const popularPost = popularPosts.map(post => {
