+++ src/pages/index.jsx
@@ -2,14 +2,14 @@ import React from "react"
 import PropTypes from "prop-types"
 import { graphql } from "gatsby"
 import { GatsbySeo } from "gatsby-plugin-next-seo"
-import { useMetadata } from "../data/use-metadata"
-import { useSetting } from "../data/use-setting"
-import { useTopics } from "../data/use-topics"
+import { useMetadata } from "../data/useMetadata"
+import { useSetting } from "../data/useSetting"
+import { useTopics } from "../data/useTopics"
 import { Text } from "../ui"
-import Layout from "../components/layout"
-import PopularPost from "../components/index/PopularPost"
-import LatestPost from "../components/index/LatestPost"
-import PostCategory from "../components/index/PostCategory"
+import Layout from "../components/Layout"
+import PopularPost from "../components/Index/PopularPost"
+import LatestPost from "../components/Index/LatestPost"
+import PostCategory from "../components/Index/PostCategory"
 
 const IndexPage = ({ data }) => {
   const {
